chore: change structure of classes
