+++ deeppavlov/models/ranking/keras_siamese_model.py
@@ -18,20 +18,20 @@ import numpy as np
 from keras import losses
 from keras.optimizers import Adam
 from keras.models import Model
-from typing import List, Iterable, Union
+from typing import List
 from abc import abstractmethod
 
 from deeppavlov.core.common.log import get_logger
+from deeppavlov.models.ranking.siamese_model import SiameseModel
 from deeppavlov.core.models.keras_model import KerasModel
 
 log = get_logger(__name__)
 
 
-class SiameseKerasModel(KerasModel):
-    """The class implementing base functionality for siamese neural networks.
+class KerasSiameseModel(SiameseModel, KerasModel):
+    """The class implementing base functionality for siamese neural networks in keras.
 
     Args:
-        batch_size: A size of a batch.
         learning_rate: Learning rate.
         use_matrix: Whether to use a trainable matrix with token (word) embeddings.
         emb_matrix: An embeddings matrix to initialize an embeddings layer of a model.
@@ -41,33 +41,24 @@ class SiameseKerasModel(KerasModel):
         dynamic_batch:  Whether to use dynamic batching. If ``True``, the maximum length of a sequence for a batch
             will be equal to the maximum of all sequences lengths from this batch,
             but not higher than ``max_sequence_length``.
-        num_context_turns: A number of ``context`` turns in data samples.
         attention: Whether any attention mechanism is used in the siamese network.
         *args: Other parameters.
         **kwargs: Other parameters.
     """
 
     def __init__(self,
-                 batch_size: int,
                  learning_rate: float = 1e-3,
                  use_matrix: bool = True,
                  emb_matrix: np.ndarray = None,
                  max_sequence_length: int = None,
                  dynamic_batch: bool = False,
-                 num_context_turns: int = 1,
                  attention: bool = False,
                  *args,
                  **kwargs):
 
-        self.save_path = kwargs.get('save_path', None)
-        self.load_path = kwargs.get('load_path', None)
-        mode = kwargs.get('mode', None)
+        super(KerasSiameseModel, self).__init__(*args, **kwargs)
 
-        super().__init__(save_path=self.save_path, load_path=self.load_path, mode=mode)
-
-        self.batch_size = batch_size
         self.learning_rate = learning_rate
-        self.num_context_turns = num_context_turns
         self.attention = attention
         self.use_matrix = use_matrix
         self.emb_matrix = emb_matrix
@@ -104,45 +95,6 @@ class SiameseKerasModel(KerasModel):
         if self.use_matrix:
             self.model.get_layer(name="embedding").set_weights([self.emb_matrix])
 
-    def train_on_batch(self, batch: List[List[List[np.ndarray]]], y: List[int]) -> float:
-        b = self._make_batch(batch)
-        loss = self._train_on_batch(b, y)
-        return loss
-
-    def __call__(self, batch: Iterable[List[np.ndarray]]) -> Union[np.ndarray, List[str]]:
-        y_pred = []
-        buf = []
-        j = 0
-        while True:
-            try:
-                el = next(batch)
-                j += 1
-                context = el[:self.num_context_turns]
-                responses = el[self.num_context_turns:]
-                buf += [context + [el] for el in responses]
-                if len(buf) >= self.batch_size:
-                    for i in range(len(buf) // self.batch_size):
-                        b = self._make_batch(buf[i*self.batch_size:(i+1)*self.batch_size])
-                        yp = self._predict_on_batch(b)
-                        y_pred += list(yp)
-                    lenb = len(buf) % self.batch_size
-                    if lenb != 0:
-                        buf = buf[-lenb:]
-                    else:
-                        buf = []
-            except StopIteration:
-                if j == 1:
-                    return ["Error! It is not intended to use the model in the interact mode."]
-                if len(buf) != 0:
-                    b = self._make_batch(buf)
-                    yp = self._predict_on_batch(b)
-                    y_pred += list(yp)
-                break
-        y_pred = np.asarray(y_pred)
-        if len(responses) > 1:
-            y_pred = np.reshape(y_pred, (j, len(responses)))
-        return y_pred
-
     @abstractmethod
     def create_model(self) -> Model:
         pass
@@ -170,14 +122,4 @@ class SiameseKerasModel(KerasModel):
 
     def _predict_response_on_batch(self, batch: List[np.ndarray]) -> np.ndarray:
         return self.response_model.predict_on_batch(batch)
-
-    def reset(self) -> None:
-        pass
-
-    def _make_batch(self, x: List[List[np.ndarray]]) -> List[np.ndarray]:
-        b = []
-        for i in range(len(x[0])):
-            z = [el[i] for el in x]
-            b.append(np.asarray(z))
-        return b
