+++ docs/apiref/models/ranking.rst
@@ -3,11 +3,13 @@ deeppavlov.models.ranking
 
 Ranking classes.
 
-.. automodule:: deeppavlov.models.ranking.siamese_keras_model
+.. automodule:: deeppavlov.models.ranking.keras_siamese_model
 
-.. autoclass:: deeppavlov.models.ranking.siamese_keras_model.SiameseKerasModel
+.. autoclass:: deeppavlov.models.ranking.siamese_model.SiameseModel
 
-.. autoclass:: deeppavlov.models.ranking.siamese_keras_predictor.SiameseKerasPredictor
+.. autoclass:: deeppavlov.models.ranking.keras_siamese_model.KerasSiameseModel
+
+.. autoclass:: deeppavlov.models.ranking.siamese_predictor.SiamesePredictor
 
 .. autoclass:: deeppavlov.models.ranking.bilstm_network.BiLSTMNetwork
